# swagger.yml
openapi: 3.0.0
paths:
...
  /api/staff:
    get:
      operationId: StaffController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List staff with search and role/category/department filters
      tags: &ref_4
        - staff
    post:
      operationId: StaffController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStaffDto'
      responses:
        '201':
          description: ''
      summary: Register a staff member (optionally linked to an identity user)
      tags: *ref_4
  /api/staff/{id}:
    get:
      operationId: StaffController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get staff member by id
      tags: *ref_4
    patch:
      operationId: StaffController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStaffDto'
      responses:
        '200':
          description: ''
      summary: Update a staff member
      tags: *ref_4
    delete:
      operationId: StaffController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a staff member
      tags: *ref_4
...
  /api/wards:
    get:
      operationId: WardsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List wards with search and filters
      tags: &ref_6
        - wards
    post:
      operationId: WardsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWardDto'
      responses:
        '201':
          description: ''
      summary: Create a ward
      tags: *ref_6
  /api/wards/{id}:
    get:
      operationId: WardsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get ward by id
      tags: *ref_6
    patch:
      operationId: WardsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWardDto'
      responses:
        '200':
          description: ''
      summary: Update a ward
      tags: *ref_6
    delete:
      operationId: WardsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a ward
      tags: *ref_6
...
  /api/visits:
    get:
      operationId: VisitsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List visits with pagination and status/provider filters
      tags: &ref_10
        - visits
    post:
      operationId: VisitsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisitDto'
      responses:
        '201':
          description: ''
      summary: Start a new visit
      tags: *ref_10
...
  /api/visits/{id}:
    get:
      operationId: VisitsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get visit by id
      tags: *ref_10
    patch:
      operationId: VisitsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVisitDto'
      responses:
        '200':
          description: ''
      summary: Update a visit
      tags: *ref_10
    delete:
      operationId: VisitsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a visit
      tags: *ref_10
  /api/visits/{id}/end:
    post:
      operationId: VisitsController_end
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndVisitDto'
      responses:
        '201':
          description: ''
      summary: End an ongoing visit
      tags: *ref_10
  /api/visits/{id}/cancel:
    post:
      operationId: VisitsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      summary: Cancel an ongoing visit
      tags: *ref_10
...
  /api/requests:
    get:
      operationId: RequestsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List clinical requests with filters
      tags: &ref_12
        - requests
    post:
      operationId: RequestsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestDto'
      responses:
        '201':
          description: ''
      summary: Create a request (optionally external-synced to LIS/pharmacy)
      tags: *ref_12
  /api/requests/{id}:
    get:
      operationId: RequestsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get request by id with items
      tags: *ref_12
    patch:
      operationId: RequestsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestDto'
      responses:
        '200':
          description: ''
      summary: Update an open request
      tags: *ref_12
    delete:
      operationId: RequestsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive a request
      tags: *ref_12
  /api/requests/{id}/history:
    get:
      operationId: RequestsController_getHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get request status history only (lightweight)
      tags: *ref_12
  /api/requests/{id}/transition:
    post:
      operationId: RequestsController_transition
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionRequestStatusDto'
      responses:
        '201':
          description: ''
      summary: Transition request status
      tags: *ref_12
  /api/requests/{id}/note:
    post:
      operationId: RequestsController_addNote
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRequestNoteDto'
      responses:
        '201':
          description: ''
      summary: Append a free-text note to the request timeline
      tags: *ref_12
  /api/requests/{id}/sync:
    post:
      operationId: RequestsController_sync
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequestDto'
      responses:
        '201':
          description: ''
      summary: Manually re-sync request to external system
      tags: *ref_12
  /api/requests/{id}/resend:
    post:
      operationId: RequestsController_resend
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      summary: Re-send a failed/open LAB or PRESCRIPTION request to the external system
      tags: *ref_12
...
info:
  title: EMR API
  description: >-
    Electronic Medical Record: appointments, visits, encounters, dynamic forms,
    and clinical requests
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateStaffDto:
      type: object
      properties:
        staffNumber:
          type: string
          description: Auto-generated as STF-xxx if omitted
        firstName:
          type: string
        lastName:
          type: string
        otherNames:
          type: string
        email:
          type: string
        phone:
          type: string
        hireDate:
          type: string
        roleType:
          type: string
          enum: &ref_16
            - Doctor
            - Nurse
            - Technician
            - Therapist
            - Admin
            - Support
            - Specialist
            - Finance
        category:
          type: string
          enum: &ref_17
            - Medical
            - Nursing
            - Allied Health
            - Pharmacy
            - Laboratory
            - Administrative
            - Support
            - Other
        department:
          type: string
        departmentId:
          type: string
          description: EMR department id (departments table)
        identityLocationId:
          type: string
          description: Location id in the identity service
        userId:
          type: string
          description: Linked identity-service user id (login)
        isActive:
          type: boolean
          default: true
        otherDetails:
          type: object
      required:
        - firstName
        - lastName
        - roleType
    UpdateStaffDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        otherNames:
          type: string
        email:
          type: string
        phone:
          type: string
        hireDate:
          type: string
        roleType:
          type: string
          enum: *ref_16
        category:
          type: string
          enum: *ref_17
        department:
          type: string
        departmentId:
          type: string
          description: EMR department id (departments table)
        identityLocationId:
          type: string
          description: Location id in the identity service
        userId:
          type: string
          description: Linked identity-service user id (login)
        isActive:
          type: boolean
        otherDetails:
          type: object
...
    CreateWardDto:
      type: object
      properties:
        code:
          type: string
          description: Short unique code, e.g. W3A
        name:
          type: string
          description: Ward display name
        wardType:
          type: string
          enum: &ref_19
            - GENERAL
            - PRIVATE
            - ICU
            - PEDIATRIC
            - MATERNITY
            - ISOLATION
            - EMERGENCY
            - OTHER
          default: GENERAL
        departmentId:
          type: string
          description: Owning EMR department id
        departmentType:
          type: string
          description: Friendly department type label
        description:
          type: string
        isActive:
          type: boolean
          default: true
      required:
        - code
        - name
    UpdateWardDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        wardType:
          type: string
          enum: *ref_19
        departmentId:
          type: string
        departmentType:
          type: string
        description:
          type: string
        isActive:
          type: boolean
...
    CreateVisitDto:
      type: object
      properties:
        patientId:
          type: string
        patientName:
          type: string
        visitType:
          type: string
          enum: &ref_25
            - OUTPATIENT
            - INPATIENT
            - EMERGENCY
            - HOME_VISIT
          default: OUTPATIENT
        providerId:
          type: string
        providerName:
          type: string
        locationId:
          type: string
        appointmentId:
          type: string
        startDatetime:
          type: string
      required:
        - patientId
        - visitType
    EndVisitDto:
      type: object
      properties:
        stopDatetime:
          type: string
    UpdateVisitDto:
      type: object
      properties:
        patientId:
          type: string
        patientName:
          type: string
        visitType:
          type: string
          enum: *ref_25
        status:
          type: string
          enum:
            - ONGOING
            - COMPLETED
            - CANCELLED
        providerId:
          type: string
        providerName:
          type: string
        locationId:
          type: string
        stopDatetime:
          type: string
...
    CreateRequestDto:
      type: object
      properties:
        patientId:
          type: string
        patientName:
          type: string
        encounterId:
          type: string
        visitId:
          type: string
        requestType:
          type: string
          enum: *ref_27
        priority:
          type: string
          enum: *ref_24
          default: ROUTINE
        orderingProviderId:
          type: string
        orderingProviderName:
          type: string
        diagnosis:
          type: string
        clinicalNotes:
          type: string
        requestedAt:
          type: string
        items:
          type: array
          items:
            type: string
      required:
        - patientId
        - requestType
        - items
    TransitionRequestStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - REQUESTED
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
            - REJECTED
        reason:
          type: string
      required:
        - status
    AddRequestNoteDto:
      type: object
      properties:
        note:
          type: string
      required:
        - note
    SyncRequestDto:
      type: object
      properties:
        externalOrderId:
          type: string
        externalReference:
          type: string
    UpdateRequestDto:
      type: object
      properties:
        patientId:
          type: string
        patientName:
          type: string
        encounterId:
          type: string
        visitId:
          type: string
        priority:
          type: string
          enum: *ref_24
        orderingProviderId:
          type: string
        orderingProviderName:
          type: string
        diagnosis:
          type: string
        clinicalNotes:
          type: string
        items:
          type: array
          items:
            type: string
...